requirements.txt
aws-cdk-lib==2.151.0
constructs==10.3.0
boto3==1.34.162
